+++ models/spring-ai-bedrock/src/main/java/org/springframework/ai/bedrock/cohere/BedrockCohereChatOptions.java
@@ -40,13 +40,13 @@ public class BedrockCohereChatOptions implements ChatOptions {
 	 * (optional) Use a lower value to decrease randomness in the response. Defaults to
 	 * 0.7.
 	 */
-	@JsonProperty("temperature") Float temperature;
+	@JsonProperty("temperature") Double temperature;
 	/**
 	 * (optional) The maximum cumulative probability of tokens to consider when sampling.
 	 * The generative uses combined Top-k and nucleus sampling. Nucleus sampling considers
 	 * the smallest set of tokens whose probability sum is at least topP.
 	 */
-	@JsonProperty("p") Float topP;
+	@JsonProperty("p") Double topP;
 	/**
 	 * (optional) Specify the number of token choices the generative uses to generate the
 	 * next token.
@@ -89,12 +89,12 @@ public class BedrockCohereChatOptions implements ChatOptions {
 
 		private final BedrockCohereChatOptions options = new BedrockCohereChatOptions();
 
-		public Builder withTemperature(Float temperature) {
+		public Builder withTemperature(Double temperature) {
 			this.options.setTemperature(temperature);
 			return this;
 		}
 
-		public Builder withTopP(Float topP) {
+		public Builder withTopP(Double topP) {
 			this.options.setTopP(topP);
 			return this;
 		}
@@ -141,20 +141,20 @@ public class BedrockCohereChatOptions implements ChatOptions {
 	}
 
 	@Override
-	public Float getTemperature() {
+	public Double getTemperature() {
 		return this.temperature;
 	}
 
-	public void setTemperature(Float temperature) {
+	public void setTemperature(Double temperature) {
 		this.temperature = temperature;
 	}
 
 	@Override
-	public Float getTopP() {
+	public Double getTopP() {
 		return this.topP;
 	}
 
-	public void setTopP(Float topP) {
+	public void setTopP(Double topP) {
 		this.topP = topP;
 	}
 
@@ -225,13 +225,13 @@ public class BedrockCohereChatOptions implements ChatOptions {
 
 	@Override
 	@JsonIgnore
-	public Float getFrequencyPenalty() {
+	public Double getFrequencyPenalty() {
 		return null;
 	}
 
 	@Override
 	@JsonIgnore
-	public Float getPresencePenalty() {
+	public Double getPresencePenalty() {
 		return null;
 	}
